Initialize npm package
