+++ .gitignore
@@ -4,3 +4,4 @@ pkg/
 test/dummy/log/*.log
 test/dummy/tmp/
 *~
+node_modules
