+++ javascript/elements/futurism_element.js
