+++ javascript/elements/futurism_table_row.js
