+++ javascript/elements/futurism_utils.js
