+++ javascript/index.js
@@ -1,0 +1,4 @@
+import { createSubscription } from './futurism_channel'
+import { initializeElements } from './elements'
+
+export { createSubscription, initializeElements }
