+++ javascript/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "@minthesize/futurism",
+  "version": "1.0.0",
+  "description": "Lazy-load Rails partials via CableReady",
+  "main": "index.js",
+  "scripts": {
+    "test": "yarn run mocha"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/julianrubisch/futurism.git"
+  },
+  "keywords": [
+    "cable_ready",
+    "\"lazy",
+    "loading\""
+  ],
+  "author": "Julian Rubisch <julian.rubisch@hey.com>",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/julianrubisch/futurism/issues"
+  },
+  "homepage": "https://github.com/julianrubisch/futurism#readme",
+  "dependencies": {
+    "cable_ready": "^4.3.0"
+  },
+  "devDependencies": {
+    "mocha": "^8.0.1",
+    "prettier-standard": "^16.4.1"
+  }
+}
